requirements.txt
pypdf>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-genai>=1.0.0
openai>=1.0.0
playwright>=1.40.0
reportlab>=4.0.0

# Web UI
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
jinja2>=3.1.4
aiofiles>=24.1.0
